cleans preprocess/process confusion
