+++ py4DSTEM/io/__init__.py
@@ -1,7 +1,25 @@
+# classes
 from py4DSTEM.io import classes
 from py4DSTEM.io.classes import *
-from py4DSTEM.io.read import *
+
+#(
+#    Tree,
+#    Root,
+#    Metadata,
+#    Array,
+#    PointList,
+#    PointListArray
+#)
+
+# read/write
+from py4DSTEM.io.read import (
+    read,
+    import_file,
+    print_h5_tree
+)
 from py4DSTEM.io.write import save
+
+# google downloader
 from py4DSTEM.io.google_drive_downloader import download_file_from_google_drive, get_sample_data_ids
