+++ py4DSTEM/io/classes/py4dstem/__init__.py
@@ -1,11 +1,25 @@
-from py4DSTEM.io.classes.py4dstem.calibration import *
-from py4DSTEM.io.classes.py4dstem.datacube import *
-from py4DSTEM.io.classes.py4dstem.probe import *
-from py4DSTEM.io.classes.py4dstem.diffractionslice import *
-from py4DSTEM.io.classes.py4dstem.virtualdiffraction import *
-from py4DSTEM.io.classes.py4dstem.realslice import *
-from py4DSTEM.io.classes.py4dstem.virtualimage import *
-from py4DSTEM.io.classes.py4dstem.qpoints import *
-from py4DSTEM.io.classes.py4dstem.braggvectors import *
+from py4DSTEM.io.classes.py4dstem.calibration import Calibration
+from py4DSTEM.io.classes.py4dstem.datacube import DataCube
+from py4DSTEM.io.classes.py4dstem.probe import Probe
+from py4DSTEM.io.classes.py4dstem.diffractionslice import DiffractionSlice
+from py4DSTEM.io.classes.py4dstem.virtualdiffraction import VirtualDiffraction
+from py4DSTEM.io.classes.py4dstem.realslice import RealSlice
+from py4DSTEM.io.classes.py4dstem.virtualimage import VirtualImage
+from py4DSTEM.io.classes.py4dstem.qpoints import QPoints
+from py4DSTEM.io.classes.py4dstem.braggvectors import BraggVectors
+from py4DSTEM.io.classes.py4dstem.parenttree import ParentTree
 from py4DSTEM.io.classes.py4dstem.io import *
+
+__pkg_classes__ = (
+    Calibration,
+    DataCube,
+    Probe,
+    DiffractionSlice,
+    VirtualDiffraction,
+    RealSlice,
+    VirtualImage,
+    QPoints,
+    BraggVectors,
+    ParentTree
+)
