+++ py4DSTEM/io/nonnative/read_dm.py
@@ -6,7 +6,7 @@ from ncempy.io import dm
 
 from py4DSTEM import tqdmnd
 from py4DSTEM.io.classes import DataCube, Array, Metadata
-from py4DSTEM.process.utils import bin2D
+from py4DSTEM.preprocess.utils import bin2D
 
 
 def read_dm(
