+++ py4DSTEM/preprocess/__init__.py
@@ -1,3 +1,4 @@
+from py4DSTEM.preprocess.utils import *
 from py4DSTEM.preprocess.preprocess import *
 from py4DSTEM.preprocess.darkreference import *
 from py4DSTEM.preprocess.electroncount import *
