+++ py4DSTEM/preprocess/electroncount.py
@@ -8,7 +8,7 @@ import numpy as np
 from scipy import optimize
 
 from py4DSTEM.io import PointListArray
-from py4DSTEM.process.utils import get_maxima_2D, bin2D
+from py4DSTEM.preprocess.utils import get_maxima_2D, bin2D
 
 def electron_count(datacube, darkreference, Nsamples=40,
                                             thresh_bkgrnd_Nsigma=4,
