+++ py4DSTEM/preprocess/utils.py
@@ -1,8 +1,98 @@
-# Get maximal points in a 2D array
+# Preprocessing utility functions
 
 import numpy as np
 from scipy.ndimage import gaussian_filter
-from py4DSTEM.process.utils.multicorr import upsampled_correlation
+
+
+def bin2D(array, factor, dtype=np.float64):
+    """
+    Bin a 2D ndarray by binfactor.
+
+    Args:
+        array (2D numpy array):
+        factor (int): the binning factor
+        dtype (numpy dtype): datatype for binned array. default is numpy default for
+            np.zeros()
+
+    Returns:
+        the binned array
+    """
+    x, y = array.shape
+    binx, biny = x // factor, y // factor
+    xx, yy = binx * factor, biny * factor
+
+    # Make a binned array on the device
+    binned_ar = np.zeros((binx, biny), dtype=dtype)
+    array = array.astype(dtype)
+
+    # Collect pixel sums into new bins
+    for ix in range(factor):
+        for iy in range(factor):
+            binned_ar += array[0 + ix:xx + ix:factor, 0 + iy:yy + iy:factor]
+    return binned_ar
+
+
+def get_shifted_ar(
+    ar,
+    xshift,
+    yshift,
+    periodic=True,
+    bilinear=False,
+    ):
+    """
+	Shifts array ar by the shift vector (xshift,yshift), using the either
+    the Fourier shift theorem (i.e. with sinc interpolation), or bilinear
+    resampling. Boundary conditions can be periodic or not.
+
+    Args:
+            ar (float): input array
+            xshift (float): shift along axis 0 (x) in pixels
+            yshift (float): shift along axis 1 (y) in pixels
+            periodic (bool): flag for periodic boundary conditions
+            bilinear (bool): flag for bilinear image shifts
+
+        Returns:
+            (array) the shifted array
+    """
+
+    # Apply image shift
+    if bilinear is False:
+        nx, ny = np.shape(ar)
+        qx, qy = make_Fourier_coords2D(nx, ny, 1)
+        nx, ny = float(nx), float(ny)
+
+        w = np.exp(-(2j * np.pi) * ((yshift * qy) + (xshift * qx)))
+        shifted_ar = np.real(np.fft.ifft2((np.fft.fft2(ar)) * w))
+
+    else:
+        xF = (np.floor(xshift)).astype(int)
+        yF = (np.floor(yshift)).astype(int)
+        wx = xshift - xF
+        wy = yshift - yF
+
+        shifted_ar = \
+            np.roll(ar,(xF  ,yF  ),axis=(0,1)) * ((1-wx)*(1-wy)) + \
+            np.roll(ar,(xF+1,yF  ),axis=(0,1)) * ((  wx)*(1-wy)) + \
+            np.roll(ar,(xF  ,yF+1),axis=(0,1)) * ((1-wx)*(  wy)) + \
+            np.roll(ar,(xF+1,yF+1),axis=(0,1)) * ((  wx)*(  wy))
+
+    if periodic is False:
+        # Rounded coordinates for boundaries
+        xR = (np.round(xshift)).astype(int)
+        yR = (np.round(yshift)).astype(int)
+
+        if xR > 0:
+            shifted_ar[0:xR,:] = 0
+        elif xR < 0:
+            shifted_ar[xR:,:] = 0
+        if yR > 0:
+            shifted_ar[:,0:yR] = 0
+        elif yR < 0:
+            shifted_ar[:,yR:] = 0
+
+    return shifted_ar
+
+
 
 
 def get_maxima_2D(
@@ -41,6 +131,8 @@ def get_maxima_2D(
     Returns:
         a structured array with fields 'x','y','intensity'
     """
+    from py4DSTEM.process.utils.multicorr import upsampled_correlation
+
     subpixel_modes = (
         'pixel',
         'poly',
@@ -126,7 +218,7 @@ def get_maxima_2D(
         subShift = upsampled_correlation(_ar_FT,upsample_factor,xyShift)
         maxima['x'][ipeak]=subShift[0]
         maxima['y'][ipeak]=subShift[1]
-        
+
     maxima = np.sort(maxima, order='intensity')[::-1]
     return maxima
 
@@ -198,6 +290,9 @@ def linear_interpolation_2D(ar, x, y):
     dy = y - y0
     return (1 - dx) * (1 - dy) * ar[x0, y0] + (1 - dx) * dy * ar[x0, y1] + dx * (1 - dy) * ar[x1, y0] + dx * dy * ar[
         x1, y1]
+
+
+
