+++ py4DSTEM/process/diskdetection/diskdetection.py
@@ -5,7 +5,7 @@ import numpy as np
 from scipy.ndimage import gaussian_filter
 
 from py4DSTEM.io.classes.py4dstem import DataCube, QPoints, BraggVectors
-from py4DSTEM.process.utils.get_maxima_2D import get_maxima_2D
+from py4DSTEM.preprocess.utils import get_maxima_2D
 from py4DSTEM.process.utils.cross_correlate import get_cross_correlation_FT
 from py4DSTEM.utils.tqdmnd import tqdmnd
 from py4DSTEM.process.diskdetection.diskdetection_aiml import find_Bragg_disks_aiml
