+++ py4DSTEM/process/utils/__init__.py
@@ -1,8 +1,16 @@
 from py4DSTEM.process.utils.utils import *
-from py4DSTEM.process.utils.get_maxima_2D import *
 from py4DSTEM.process.utils.cross_correlate import *
 from py4DSTEM.process.utils.multicorr import *
 from py4DSTEM.process.utils.elliptical_coords import *
 from py4DSTEM.process.utils.masks import *
 from py4DSTEM.process.utils.single_atom_scatter import *
+
+# from preprocessing
+from py4DSTEM.preprocess.utils import (
+    bin2D,
+    get_maxima_2D,
+    get_shifted_ar,
+    filter_2D_maxima,
+    linear_interpolation_2D
+)
